requirements.txt
arxiv>=2.1.0
PyYAML>=6.0
requests>=2.31.0
